Ansible/requirements.txt
huaweicloudsdkcore==3.1.107
huaweicloudsdkecs==3.1.107
huaweicloudsdkevs==3.1.107
python-dotenv==1.0.1
PyYAML==6.0.1
PyYAML==6.0.1
Requests==2.32.3
